PAL-616-Duplicated-SimpleClient-in-Examples-And-Clients-Repo: Moved related classes to the Client Repository.
