+++ example-model/src/main/java/uk/gov/gchq/palisade/example/config/ApplicationConfiguration.java
@@ -28,12 +28,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 
+import uk.gov.gchq.palisade.clients.simpleclient.web.DataClient;
+import uk.gov.gchq.palisade.clients.simpleclient.web.PalisadeClient;
 import uk.gov.gchq.palisade.example.client.ExampleSimpleClient;
 import uk.gov.gchq.palisade.example.configurator.ExampleConfigurator;
 import uk.gov.gchq.palisade.example.runner.BulkTestExample;
 import uk.gov.gchq.palisade.example.runner.RestExample;
-import uk.gov.gchq.palisade.example.web.DataClient;
-import uk.gov.gchq.palisade.example.web.PalisadeClient;
 import uk.gov.gchq.palisade.example.web.PolicyClient;
 import uk.gov.gchq.palisade.example.web.ResourceClient;
 import uk.gov.gchq.palisade.example.web.UserClient;
