+++ example-model/src/main/java/uk/gov/gchq/palisade/example/configurator/ExampleConfigurator.java
@@ -22,11 +22,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.cloud.client.ServiceInstance;
 import org.springframework.cloud.netflix.eureka.EurekaServiceInstance;
 
+import uk.gov.gchq.palisade.clients.simpleclient.request.AddSerialiserRequest;
+import uk.gov.gchq.palisade.clients.simpleclient.web.DataClient;
 import uk.gov.gchq.palisade.data.serialise.AvroSerialiser;
 import uk.gov.gchq.palisade.example.hrdatagenerator.types.Employee;
-import uk.gov.gchq.palisade.example.request.AddSerialiserRequest;
 import uk.gov.gchq.palisade.example.util.ExampleFileUtil;
-import uk.gov.gchq.palisade.example.web.DataClient;
 import uk.gov.gchq.palisade.example.web.PolicyClient;
 import uk.gov.gchq.palisade.example.web.ResourceClient;
 import uk.gov.gchq.palisade.example.web.UserClient;
